Re-worked bass crossover to remove bass timing delay
Crossover was adding 2-3ms of delay to the bass frequencies because of
the crossover. I have removed the crossover from the original signal and
attenuated the bass a bit using filters. I then adjusted the delay on
the bass signals to adjust for the fact that the crossover adds delay.
Made some small EQ tweaks to get freqency response cleaner.